fixed linters
