+++ teacher.rb
@@ -1,6 +1,8 @@
 require './person'
 
 class Teacher < Person
+  attr_accessor :specialization
+
   def initialize(age, specialization, name = 'unknown')
     super(age, name)
     @specialization = specialization
