+++ trimmer_decorator.rb
@@ -7,6 +7,6 @@ class TrimmerDecorator < Nameable
   end
 
   def correct_name
-    @nameable.correct_name.length > 10 ? @nameable.correct_name[0..9] : @nameable.correct_name
+    @nameable.correct_name.slice(0, 10)
   end
 end
